Add a script to report outdated module usage
Outputs a JSON document with a list of every outdated (i.e. not the
latest version) terraform module used. The report includes the namespace
name, module name, version in use and latest version.

The idea is to use this to populate a page in the "how out of date are
we" application.